requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.4=py38hef030d1_0
aiosignal=1.3.1=pyhd8ed1ab_0
appnope=0.1.3=pypi_0
arrow-cpp=9.0.0=py38hf6a008a_9_cpu
asttokens=2.2.1=pypi_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
aws-c-cal=0.5.11=hd2e2f4b_0
aws-c-common=0.6.2=h0d85af4_0
aws-c-event-stream=0.2.7=hb9330a7_13
aws-c-io=0.10.5=h35aa462_0
aws-checksums=0.1.11=h0010a65_7
aws-sdk-cpp=1.8.186=h7c85d8e_4
backcall=0.2.0=pypi_0
blinker=1.6.2=pyhd8ed1ab_0
boltons=23.0.0=pyhd8ed1ab_0
brotlipy=0.7.0=py38hef030d1_1005
bzip2=1.0.8=h0d85af4_4
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.12.7=h033912b_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.0=pyhd8ed1ab_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.14.4=py38h6afc60c_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.3=pypi_0
conda=23.3.1=py38h50d1736_0
conda-package-handling=2.0.2=pyh38be061_0
conda-package-streaming=0.7.0=pyhd8ed1ab_1
contourpy=1.0.7=pypi_0
cryptography=39.0.0=py38ha6c3189_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.11.0=pyhd8ed1ab_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.6=pyhd8ed1ab_1
executing=1.2.0=pypi_0
filelock=3.12.0=pyhd8ed1ab_0
flatbuffers=2.0.8=hf0c8a7f_1
fonttools=4.39.3=pypi_0
frozenlist=1.3.3=py38hef030d1_0
fsspec=2023.4.0=pyh1a96a4e_0
gast=0.4.0=pyh9f0ad1d_0
gensim=4.3.1=py38hdaefe69_1
gflags=2.2.2=hb1e8313_1004
giflib=5.2.1=hb7f2c08_3
glog=0.6.0=h8ac2a54_0
google-auth=2.17.3=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpc-cpp=1.47.1=h834a566_7
grpcio=1.47.1=py38h365929b_7
h5py=3.8.0=nompi_py38hd2287a7_101
hdf5=1.14.0=nompi_h058c35b_103
huggingface_hub=0.13.4=pyhd8ed1ab_0
icu=70.1=h96cf925_0
idna=3.4=pyhd8ed1ab_0
imageio=2.27.0=pypi_0
importlib-metadata=6.5.0=pyha770c72_0
importlib-resources=5.12.0=pypi_0
importlib_metadata=6.5.0=hd8ed1ab_0
ipykernel=6.22.0=pypi_0
ipython=8.12.0=pypi_0
ipywidgets=8.0.6=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=hb7f2c08_3
jsonpatch=1.32=pyhd8ed1ab_0
jsonpointer=2.0=py_0
jupyter-client=8.2.0=pypi_0
jupyter-core=5.3.0=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
keras=2.10.0=pyhd8ed1ab_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=h0165f36_0
lazy-loader=0.2=pypi_0
libabseil=20220623.0=cxx17_h844d122_6
libaec=1.0.6=hf0c8a7f_1
libblas=3.9.0=16_osx64_openblas
libbrotlicommon=1.0.9=hb7f2c08_8
libbrotlidec=1.0.9=hb7f2c08_8
libbrotlienc=1.0.9=hb7f2c08_8
libcblas=3.9.0=16_osx64_openblas
libcrc32c=1.1.2=he49afe7_0
libcurl=7.87.0=haf73cf8_0
libcxx=16.0.1=h71dddab_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libevent=2.1.10=h815e4d9_4
libffi=3.2.1=hb1e8313_1007
libgfortran=5.0.0=11_3_0_h97931a8_31
libgfortran5=12.2.0=he409387_31
libgoogle-cloud=2.3.0=hb0fe3b0_0
liblapack=3.9.0=16_osx64_openblas
libnghttp2=1.51.0=h0dd9d14_0
libopenblas=0.3.21=openmp_h429af6e_3
libpng=1.6.39=ha978bb4_0
libprotobuf=3.21.12=hbc0c0cd_0
libsqlite=3.40.0=ha978bb4_0
libssh2=1.10.0=h7535e13_3
libthrift=0.16.0=h08c06f4_2
libutf8proc=2.8.0=hb7f2c08_0
libzlib=1.2.13=hfd90126_4
llvm-openmp=16.0.1=h61d9ccf_0
lz4-c=1.9.4=hf0c8a7f_0
markdown=3.4.3=pyhd8ed1ab_0
markupsafe=2.1.2=py38hef030d1_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mediapipe=0.9.1.0=pypi_0
mkl=2022.2.1=h44ed08c_16952
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.0.4=py38hef030d1_0
multiprocess=0.70.14=py38hef030d1_3
ncurses=6.3=h96cf925_1
nest-asyncio=1.5.6=pypi_0
networkx=3.1=pyhd8ed1ab_0
nltk=3.8.1=pyhd8ed1ab_0
numpy=1.24.2=py38h5a2dcdf_0
oauthlib=3.2.2=pyhd8ed1ab_0
opencv-contrib-python=4.7.0.72=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=hfd90126_0
opt_einsum=3.3.0=pyhd8ed1ab_1
orc=1.8.0=ha9d861c_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.0=py38hec72209_0
parquet-cpp=1.5.1=2
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.1=pyhd8ed1ab_0
platformdirs=3.2.0=pyhd8ed1ab_0
pluggy=1.0.0=pyhd8ed1ab_5
pooch=1.7.0=pyha770c72_3
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=9.0.0=py38hf59b08f_9_cpu
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycosat=0.6.4=py38hef030d1_1
pycparser=2.21=pyhd8ed1ab_0
pygments=2.15.1=pypi_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=23.1.1=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.0=hd366da7_5
python-dateutil=2.8.2=pyhd8ed1ab_0
python-flatbuffers=23.1.21=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python-xxhash=3.2.0=py38hef030d1_0
python_abi=3.8=2_cp38
pytorch=2.0.0=cpu_py38h7c2c2ba_0
pytz=2023.3=pyhd8ed1ab_0
pyu2f=0.1.5=pyhd8ed1ab_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=py38hef030d1_5
pyzmq=25.0.2=pypi_0
re2=2022.06.01=hb486fe8_1
readline=8.2=h9e318b2_1
regex=2023.3.23=py38hef030d1_0
requests=2.28.2=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
responses=0.18.0=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
ruamel.yaml=0.17.21=py38hef030d1_3
ruamel.yaml.clib=0.2.7=py38hef030d1_1
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-image=0.20.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.9.1=pypi_0
setuptools=67.7.0=pypi_0
six=1.16.0=pyh6c4a22f_0
sklearn=0.0.post4=pypi_0
sleef=3.5.1=h6db0672_2
smart_open=6.3.0=pyhd8ed1ab_1
snappy=1.1.10=h225ccf5_0
sqlite=3.40.0=h9ae0607_0
stack-data=0.6.2=pypi_0
sympy=1.11.1=pyh04b8f61_3
tbb=2021.9.0=hb8565cd_0
tensorboard=2.10.1=pyhd8ed1ab_0
tensorboard-data-server=0.6.1=py38ha6c3189_4
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.10.0=cpu_py38he866dac_0
tensorflow-base=2.10.0=cpu_py38h13e9d53_0
tensorflow-estimator=2.10.0=cpu_py38h7c3e1b3_0
termcolor=2.2.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.4.12=pypi_0
tk=8.6.12=h5dbffcc_0
tokenizers=0.13.3=py38hc426687_0
toolz=0.12.0=pyhd8ed1ab_0
torch-geometric=2.3.0=pypi_0
torchvision=0.15.1=pypi_0
tornado=6.3.1=pypi_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pypi_0
transformers=4.28.1=pyhd8ed1ab_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
urllib3=1.26.15=pyhd8ed1ab_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.3=pyhd8ed1ab_0
wheel=0.40.0=pyhd8ed1ab_0
widgetsnbextension=4.0.7=pypi_0
wrapt=1.15.0=py38hef030d1_0
xxhash=0.8.1=h45a561f_0
xz=5.2.6=h775f41a_0
yaml=0.2.5=h0d85af4_2
yarl=1.8.2=py38hef030d1_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=hfd90126_4
zstandard=0.19.0=py38h7cbeef7_1
zstd=1.5.2=hbc0c0cd_6
